access token and refresh token implement

+++ routes/authRoutes.js
@@ -1,14 +1,20 @@
 const express = require('express');
-const { createUser, getAllUser, loginUser, deleteUser, deleteAllUser, updateUserByEmail, getuser } = require('../controller/userCtrl');
+const { createUser, getAllUser, loginUser, deleteUser, deleteAllUser, updateUserByEmail, getuser, blockUser, unblockUser, updateUser, getAccessToken, logoutUser } = require('../controller/userCtrl');
 const { authMiddleware, isAdmin } = require('../middlewares/authMiddleware');
 const router = express.Router();
 
 router.post('/register', createUser);	
-router.get('/getUser/:id', authMiddleware, isAdmin, getuser);
+router.get('/getUser', authMiddleware, getuser);
 router.get('/getAllUser', getAllUser);
 router.get('/login', loginUser);
-router.delete('/delete', deleteUser);
-router.delete('/deleteAllUser', deleteAllUser);
-router.put('/updateUserByEmail/:email',authMiddleware, updateUserByEmail)
+router.get('/getAccessToken', getAccessToken);
+router.delete('/delete', deleteUser); // todo: only admin can delete
+router.delete('/deleteAllUser', deleteAllUser); // todo: only admin can delete
+router.put('/update-user', authMiddleware, updateUser);
+router.put('/updateUserByEmail',authMiddleware, updateUserByEmail)
+router.put('/block-user/:email', authMiddleware, isAdmin, blockUser);
+router.put('/unblock-user/:email',authMiddleware, isAdmin, unblockUser);
+router.put('/logout',authMiddleware, logoutUser);
+
 
 module.exports = router;
